Develop into master

+++ src/main/java/org/timowa/megabazar/dto/user/UserProfileDto.java
@@ -19,5 +19,5 @@ public class UserProfileDto {
 
     private LocalDateTime createdAt;
 
-    private Integer money;
+    private Double money;
 }
